# tests/contract/snapshots/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "AnswerStyle": {
        "enum": [
          "concise",
          "summary"
        ],
        "title": "AnswerStyle",
        "type": "string"
      },
...
      "DualSpaceEmbeddingSelection": {
        "additionalProperties": false,
        "properties": {
          "multimodal_profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multimodal Profile Revision Id"
          },
          "strategy": {
            "const": "dual_space",
            "default": "dual_space",
            "title": "Strategy",
            "type": "string"
          },
          "text_profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Profile Revision Id"
          }
        },
        "title": "DualSpaceEmbeddingSelection",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "REQUEST_VALIDATION_FAILED",
          "INVALID_CURSOR",
          "INVALID_IDEMPOTENCY_KEY",
          "IDEMPOTENCY_KEY_REUSED",
          "CAPABILITY_NOT_ENABLED",
          "ANSWER_POLICY_NOT_SUPPORTED",
          "CHAT_SESSION_BUSY",
          "CHAT_CONTEXT_INVALID",
          "CHAT_REVISION_MISMATCH",
          "CHAT_PROVIDER_UNAVAILABLE",
          "CHAT_RESPONSE_INVALID",
          "CHAT_PIPELINE_DEADLINE_EXCEEDED",
          "CHAT_PIPELINE_STEP_FAILED",
          "CHAT_PERSISTENCE_FAILED",
          "CHAT_STALE_WORKER",
          "CHAT_WORKER_STOPPED",
          "CHAT_RUN_CANCELLED",
          "CHAT_SSE_CONNECTION_LIMIT_EXCEEDED",
          "IDENTITY_OVERRIDE_NOT_ALLOWED",
          "ACCESS_DENIED",
          "RESOURCE_NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "HTTP_ERROR",
          "INTERNAL_SERVER_ERROR",
          "RESOURCE_NAME_CONFLICT",
          "RESOURCE_STATE_CONFLICT",
          "FILE_NAME_INVALID",
          "FILE_MEDIA_TYPE_UNSUPPORTED",
          "FILE_MEDIA_TYPE_MISMATCH",
          "FILE_TOO_LARGE",
          "FILE_INVALID_UTF8",
          "FILE_LINE_LIMIT_EXCEEDED",
          "FILE_STRUCTURE_LIMIT_EXCEEDED",
          "FILE_CONTENT_INVALID",
          "FILE_ARCHIVE_LIMIT_EXCEEDED",
          "MARKDOWN_BUNDLE_INVALID",
          "MARKDOWN_MEDIA_UNRESOLVED",
          "MARKDOWN_MEDIA_UNSUPPORTED",
          "MARKDOWN_MEDIA_FETCH_FAILED",
          "PARSER_NOT_CONFIGURED",
          "PARSER_CHUNK_LIMIT_EXCEEDED",
          "PARSER_RESOURCE_LIMIT",
          "PARSER_CRASHED",
          "PARSER_OUTPUT_INVALID",
          "INDEX_TARGET_INVALID",
          "INDEX_REVISION_INCOMPATIBLE",
          "INDEX_CHUNK_PLAN_MISMATCH",
          "SEMANTIC_CHUNKING_FAILED",
          "SOURCE_NOT_AVAILABLE",
          "SOURCE_FILE_MISSING",
          "SOURCE_FILE_INTEGRITY",
          "EMBEDDING_SPACE_MISMATCH",
          "EMBEDDING_PROVIDER_UNAVAILABLE",
          "EMBEDDING_RESPONSE_INVALID",
          "LOCAL_RERANKER_UNAVAILABLE",
          "RETRIEVAL_DEADLINE_EXCEEDED",
          "GRAPH_NOT_READY",
          "GRAPH_CONFIG_INVALID",
          "GRAPH_PROVIDER_UNAVAILABLE",
          "GRAPH_PROTOCOL_INVALID",
          "GRAPH_RESOURCE_LIMIT",
          "GRAPH_BUILD_FAILED",
          "INDEX_PERSISTENCE_FAILED",
          "INDEX_INCOMPLETE",
          "INDEXING_STALE_WORKER",
          "INDEXING_WORKER_STOPPED"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "FieldViolation": {
        "additionalProperties": false,
        "properties": {
          "error_type": {
            "title": "Error Type",
            "type": "string"
          },
          "location": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "location",
          "message",
          "error_type"
        ],
        "title": "FieldViolation",
        "type": "object"
      },
...
      "InsufficiencyPolicy": {
        "enum": [
          "refuse",
          "partial_answer"
        ],
        "title": "InsufficiencyPolicy",
        "type": "string"
      },
      "KnowledgeBaseAnswerPolicyDefaults": {
        "additionalProperties": false,
        "properties": {
          "answer_style": {
            "$ref": "#/components/schemas/AnswerStyle",
            "default": "concise"
          },
          "insufficiency_policy": {
            "$ref": "#/components/schemas/InsufficiencyPolicy",
            "default": "partial_answer"
          }
        },
        "title": "KnowledgeBaseAnswerPolicyDefaults",
        "type": "object"
      },
      "KnowledgeBaseChunking": {
        "additionalProperties": false,
        "properties": {
          "preset": {
            "$ref": "#/components/schemas/ChunkingPreset",
            "default": "structural_balanced_v2"
          }
        },
        "title": "KnowledgeBaseChunking",
        "type": "object"
      },
      "KnowledgeBaseChunkingResponse": {
        "additionalProperties": false,
        "properties": {
          "preset": {
            "$ref": "#/components/schemas/ChunkingPreset"
          },
          "profile": {
            "enum": [
              "structural_by_title_token_v4",
              "semantic_breakpoint_v3",
              "semantic_breakpoint_v4"
            ],
            "title": "Profile",
            "type": "string"
          }
        },
        "required": [
          "preset",
          "profile"
        ],
        "title": "KnowledgeBaseChunkingResponse",
        "type": "object"
      },
      "KnowledgeBaseCreate": {
        "additionalProperties": false,
        "properties": {
          "answer_policy_defaults": {
            "$ref": "#/components/schemas/KnowledgeBaseAnswerPolicyDefaults",
            "default": {
              "answer_style": "concise",
              "insufficiency_policy": "partial_answer"
            }
          },
          "chunking": {
            "$ref": "#/components/schemas/KnowledgeBaseChunking",
            "default": {
              "preset": "structural_balanced_v2"
            }
          },
          "embedding": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "dual_space": "#/components/schemas/DualSpaceEmbeddingSelection",
                    "text_only": "#/components/schemas/TextOnlyEmbeddingSelection",
                    "unified_multimodal": "#/components/schemas/UnifiedMultimodalEmbeddingSelection"
                  },
                  "propertyName": "strategy"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/TextOnlyEmbeddingSelection"
                  },
                  {
                    "$ref": "#/components/schemas/DualSpaceEmbeddingSelection"
                  },
                  {
                    "$ref": "#/components/schemas/UnifiedMultimodalEmbeddingSelection"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "parsing": {
            "$ref": "#/components/schemas/KnowledgeBaseParsing",
            "default": {
              "preset": "text_local_v1"
            }
          },
          "retrieval_defaults": {
            "$ref": "#/components/schemas/RetrievalDefaults",
            "default": {
              "rerank_mode": "classic",
              "strategy": "exact_vector",
              "top_k": 10
            }
          }
        },
        "required": [
          "name"
        ],
        "title": "KnowledgeBaseCreate",
        "type": "object"
      },
      "KnowledgeBaseDeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "deleted_at": {
            "format": "date-time",
            "title": "Deleted At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "deleted_at"
        ],
        "title": "KnowledgeBaseDeleteResponse",
        "type": "object"
      },
      "KnowledgeBaseEmbeddingResponse": {
        "additionalProperties": false,
        "properties": {
          "cross_modal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeBaseEmbeddingRoleResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "strategy": {
            "enum": [
              "text_only",
              "dual_space",
              "unified_multimodal"
            ],
            "title": "Strategy",
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/KnowledgeBaseEmbeddingRoleResponse"
          }
        },
        "required": [
          "strategy",
          "text"
        ],
        "title": "KnowledgeBaseEmbeddingResponse",
        "type": "object"
      },
...
      "KnowledgeBasePage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeBaseResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "KnowledgeBasePage",
        "type": "object"
      },
      "KnowledgeBaseParsing": {
        "additionalProperties": false,
        "properties": {
          "preset": {
            "$ref": "#/components/schemas/ParsingPreset",
            "default": "text_local_v1"
          }
        },
        "title": "KnowledgeBaseParsing",
        "type": "object"
      },
...
      "KnowledgeBaseResponse": {
        "additionalProperties": false,
        "properties": {
          "active_index_revision_id": {
            "format": "uuid",
            "title": "Active Index Revision Id",
            "type": "string"
          },
          "answer_policy_defaults": {
            "$ref": "#/components/schemas/KnowledgeBaseAnswerPolicyDefaults",
            "default": {
              "answer_style": "concise",
              "insufficiency_policy": "partial_answer"
            }
          },
          "chunking": {
            "$ref": "#/components/schemas/KnowledgeBaseChunkingResponse"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "embedding": {
            "$ref": "#/components/schemas/KnowledgeBaseEmbeddingResponse"
          },
          "embedding_space_id": {
            "format": "uuid",
            "title": "Embedding Space Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parsing": {
            "$ref": "#/components/schemas/KnowledgeBaseParsingResponse"
          },
          "provisioned_at": {
            "format": "date-time",
            "title": "Provisioned At",
            "type": "string"
          },
          "retrieval_defaults": {
            "$ref": "#/components/schemas/RetrievalDefaults"
          },
          "source_change_seq": {
            "title": "Source Change Seq",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "source_change_seq",
          "active_index_revision_id",
          "embedding_space_id",
          "embedding",
          "parsing",
          "chunking",
          "retrieval_defaults",
          "provisioned_at",
          "created_at",
          "updated_at"
        ],
        "title": "KnowledgeBaseResponse",
        "type": "object"
      },
      "KnowledgeBaseUpdate": {
        "additionalProperties": false,
        "properties": {
          "answer_policy_defaults": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KnowledgeBaseAnswerPolicyDefaults"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "retrieval_defaults": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalDefaults"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "KnowledgeBaseUpdate",
        "type": "object"
      },
...
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldViolation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "title": "Instance",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "trace_id",
          "retryable"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "RetrievalDefaults": {
        "additionalProperties": false,
        "properties": {
          "rerank_mode": {
            "$ref": "#/components/schemas/RerankMode",
            "default": "classic"
          },
          "strategy": {
            "const": "exact_vector",
            "default": "exact_vector",
            "title": "Strategy",
            "type": "string"
          },
          "top_k": {
            "default": 10,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "title": "RetrievalDefaults",
        "type": "object"
      },
...
      "TextOnlyEmbeddingSelection": {
        "additionalProperties": false,
        "properties": {
          "strategy": {
            "const": "text_only",
            "default": "text_only",
            "title": "Strategy",
            "type": "string"
          },
          "text_profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text Profile Revision Id"
          }
        },
        "title": "TextOnlyEmbeddingSelection",
        "type": "object"
      },
      "UnifiedMultimodalEmbeddingSelection": {
        "additionalProperties": false,
        "properties": {
          "profile_revision_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Revision Id"
          },
          "strategy": {
            "const": "unified_multimodal",
            "default": "unified_multimodal",
            "title": "Strategy",
            "type": "string"
          }
        },
        "title": "UnifiedMultimodalEmbeddingSelection",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Enterprise Knowledge Base API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge-bases": {
      "get": {
        "operationId": "list_knowledge_bases_api_v1_knowledge_bases_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 2048,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "created_at",
              "enum": [
                "created_at",
                "-created_at",
                "name",
                "-name"
              ],
              "title": "Sort",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBasePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Invalid request or cursor"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "List Knowledge Bases",
        "tags": [
          "knowledge-bases"
        ]
      },
      "post": {
        "operationId": "create_knowledge_base_api_v1_knowledge_bases_post",
        "parameters": [
          {
            "description": "Required UUID scoped by principal, client, and endpoint.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Required UUID scoped by principal, client, and endpoint.",
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Create Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
    "/api/v1/knowledge-bases/{kb_id}": {
      "delete": {
        "operationId": "delete_knowledge_base_api_v1_knowledge_bases__kb_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "description": "Required UUID scoped by principal, client, and endpoint.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Required UUID scoped by principal, client, and endpoint.",
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseDeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Delete Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      },
      "get": {
        "operationId": "get_knowledge_base_api_v1_knowledge_bases__kb_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Get Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      },
      "patch": {
        "operationId": "update_knowledge_base_api_v1_knowledge_bases__kb_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "kb_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Kb Id",
              "type": "string"
            }
          },
          {
            "description": "Required UUID scoped by principal, client, and endpoint.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "description": "Required UUID scoped by principal, client, and endpoint.",
              "format": "uuid",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeBaseResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Idempotency or resource conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Request validation failed"
          }
        },
        "summary": "Update Knowledge Base",
        "tags": [
          "knowledge-bases"
        ]
      }
    },
...
  }
}